+++ fe/fe-core/src/main/java/com/starrocks/sql/analyzer/QueryAnalyzer.java
@@ -57,8 +57,6 @@ import com.starrocks.common.profile.Timer;
 import com.starrocks.common.profile.Tracers;
 import com.starrocks.common.util.concurrent.lock.LockType;
 import com.starrocks.common.util.concurrent.lock.Locker;
-import com.starrocks.connector.ConnectorTableVersion;
-import com.starrocks.connector.PointerType;
 import com.starrocks.privilege.SecurityPolicyRewriteRule;
 import com.starrocks.qe.ConnectContext;
 import com.starrocks.server.GlobalStateMgr;
@@ -76,7 +74,6 @@ import com.starrocks.sql.ast.PartitionNames;
 import com.starrocks.sql.ast.PivotAggregation;
 import com.starrocks.sql.ast.PivotRelation;
 import com.starrocks.sql.ast.PivotValue;
-import com.starrocks.sql.ast.QueryPeriod;
 import com.starrocks.sql.ast.QueryRelation;
 import com.starrocks.sql.ast.QueryStatement;
 import com.starrocks.sql.ast.Relation;
@@ -93,12 +90,7 @@ import com.starrocks.sql.ast.ValuesRelation;
 import com.starrocks.sql.ast.ViewRelation;
 import com.starrocks.sql.common.MetaUtils;
 import com.starrocks.sql.common.TypeManager;
-import com.starrocks.sql.optimizer.base.ColumnRefFactory;
 import com.starrocks.sql.optimizer.dump.HiveMetaStoreTableDumpInfo;
-import com.starrocks.sql.optimizer.operator.scalar.ConstantOperator;
-import com.starrocks.sql.optimizer.operator.scalar.ScalarOperator;
-import com.starrocks.sql.optimizer.transformer.ExpressionMapping;
-import com.starrocks.sql.optimizer.transformer.SqlToScalarOperatorTranslator;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -1256,16 +1248,7 @@ public class QueryAnalyzer {
                 }
                 if (table == null) {
                     try (Timer ignored = Tracers.watchScope("AnalyzeTable")) {
-                        if (tableRelation.getQueryPeriod() == null) {
-                            table = metadataMgr.getTable(catalogName, dbName, tbName);
-                        } else {
-                            // for time travel
-                            QueryPeriod queryPeriod = tableRelation.getQueryPeriod();
-                            QueryPeriod.PeriodType periodType = queryPeriod.getPeriodType();
-                            Optional<ConnectorTableVersion> startVersion = resolveQueryPeriod(queryPeriod.getStart(), periodType);
-                            Optional<ConnectorTableVersion> endVersion = resolveQueryPeriod(queryPeriod.getEnd(), periodType);
-                            table = metadataMgr.getTable(catalogName, dbName, tbName, startVersion, endVersion);
-                        }
+                        table = metadataMgr.getTable(catalogName, dbName, tbName);
                     }
                 }
             }
@@ -1303,34 +1286,6 @@ public class QueryAnalyzer {
         } catch (AnalysisException e) {
             throw new SemanticException(e.getMessage());
         }
-    }
-
-    private Optional<ConnectorTableVersion> resolveQueryPeriod(Optional<Expr> version, QueryPeriod.PeriodType type) {
-        if (version.isEmpty()) {
-            return Optional.empty();
-        }
-        ScalarOperator result;
-        try {
-            Scope scope = new Scope(RelationId.anonymous(), new RelationFields());
-            ExpressionAnalyzer.analyzeExpression(version.get(), new AnalyzeState(), scope, session);
-            ExpressionMapping expressionMapping = new ExpressionMapping(scope);
-            result = SqlToScalarOperatorTranslator.translate(version.get(), expressionMapping, new ColumnRefFactory());
-        } catch (Exception e) {
-            throw new SemanticException("Failed to resolve query period [type: %s, value: %s]. msg: %s",
-                    type.toString(), version.get().toString(), e.getMessage());
-        }
-
-        if (!(result instanceof ConstantOperator)) {
-            if (version.get() instanceof FunctionCallExpr) {
-                throw new SemanticException("Invalid datetime function: [type: %s, value: %s]. " +
-                        "The function requirement must be inferred in frontend.", type.toString(), version.get().toString());
-            } else {
-                throw new SemanticException("Invalid version value. [type: %s, value: %s]",
-                        type.toString(), version.get().toString());
-            }
-        }
-        PointerType pointerType = type == QueryPeriod.PeriodType.TIMESTAMP ? PointerType.TEMPORAL : PointerType.VERSION;
-        return Optional.of(new ConnectorTableVersion(pointerType, (ConstantOperator) result));
     }
 
     private Table resolveTableFunctionTable(Map<String, String> properties) {
